change interface administrator -> view, edit and delete for any product, only worker edit with product type procesor

+++ Common/Persona.cs
@@ -8,7 +8,7 @@ namespace Univalle.Fie.Sistemas.BaseDeDatos2.AppComputadorasBDD.Common
         public Guid IdPersona { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public short Sexo { get; set; }
+        public byte Sexo { get; set; }
         public bool Eliminado { get; set; }
         public Usuario Usuario { get; set; }
         public DateTime FechaNacimiento { get; set; }
